+++ helper/helper.go
@@ -90,7 +90,7 @@ type HeaderValue struct {
 // then
 //  [{text/html 1} {*/* 0.7} {text/* 0.5}]
 func ParseMultiValuesHeader(header string) []HeaderValue {
-	regex := regexp.MustCompile("^q=([01]\\.[0-9]{1,3})$")
+	regex := regexp.MustCompile(`^q=([01]\.[0-9]{1,3})$`)
 	split := strings.Split(header, ",")
 	values := make([]HeaderValue, 0, len(split))