+++ router_test.go
@@ -298,7 +298,6 @@ func (suite *RouterTestSuite) TestStatusHandlers() {
 	// On subrouters
 	subrouter := router.Subrouter("/sub")
 	writer = httptest.NewRecorder()
-	router = newRouter()
 
 	subrouter.requestHandler(match, writer, rawRequest)